separate OTUnitTest
